linux-ramdump-parser-v2: Make the page print outs more useful
The pagetype info gives useful data but it's helpful to see the data
in other formats that often involve having to do math. Make the parser
do the math of calculating the total number of pages per migrate type
and the total number of pages per order. Additionally, print out the
total number of ram pages with vmstat to make it easier to determine
just how low a system may be on memory.

Change-Id: I812288fe4c25de06c7017fb0b6515e514f799134