linux-ramdump-parser-v2: Improve tracing for socinfo and other parts
When adding a new board, it can be non-obvious what went wrong if the
board is not immediately detected. Add an option to trace through
socinfo addresses for a specific board to print out useful addresses
as a debugging aid. Also clean up some error handling, mostly among
the tracing code.

Change-Id: I65f4a38b95912b81676f44cd371751abc96aae9e