linux-ramdump-parser-v2: Conditionally add phys offset for socinfo
Commit 3d6a90bd7106e9e8dd1718bad4e35ea77ffd802b
(linux-ramdump-parser-v2: Don't use ram start address in socinfo)
corrected the case for --auto-dump detection but broke the case of
specifying ram files manually with addresses. When specifying the
addresses manually, the addresses given are correct so the ram start
does need to be added. Correct the code to add the offset by default
but not add when doing autodetection.

Change-Id: I50d54a932210afeace49a898dd819dc46c3e3aa1