lrdp-v2: add interactive shell
During development, it's quite convenient to load up an interactive
python interpreter with a RamDump instance loaded. Add some command line
switches to launch an interactive shell with the RamDump instance that
gets created as a result of whatever other command line options are
passed.

If IPython is installed, an IPython interpreter will be launched,
otherwise a classic Python interpreter will be launched. The classic
interface can also be forced with --classic-shell.

Change-Id: Id59acb97a830055212de9db3eaf05f18358f757f