lrdp-v2: improve handling of unsupported python versions
We don't support python3 at all and python2.6 can only be used if the
user installs an extra package (OrderedDict [1]). Ask for python2
explicitly in the shebang line to resolve the python3 issues and print
some instructions to the user regarding python2.6 when that is used.

If the user insists on using python2.6, provide a command line
switch (--force-26) to skip the version check. To reduce clutter, don't
document this switch in the --help text but inform the user about its
existence when we error out due to the python2.6 check.

In order to even print an error message in python3 we have to first
ensure that our code is valid in python3, otherwise the interpreter
fails to even start executing our code (so we can't print any
messages). Use 2to3 to patch up some print statements, etc. Also fix
inconsistent whitespace, which is a syntax error on python3.

[1] https://pypi.python.org/pypi/ordereddict

Change-Id: Ie2c0a200e60ec90bf6cf49789f2cc75f181fa94b