lrdp-v2: sort the parsers before running
Currently the registered parsers can be run in basically any order. It
would be slightly messy to run them in the order the user specified on
the command line, so just run them alphabetically for now.

This is needed to do regression testing on the LRDP (diff'ing output of
the same dumps with different LRDP code).

Change-Id: Ib63c9b5fecaf1170df0e9e121980fcd0e4f35623