lrdpv2: use explicitly-sized types in Iommu parsing
The `RamParse.read_word' function reads a different number of bits
depending on the architecture (it reads 32 bits on ARM and 64 bits on
ARM64). We were relying on it to only read 32 bits in a few places, so
things were broken on ARM64. Fix this by using explicitly-sized reads
where appropriate.

Change-Id: Id717787c4e7df344964387d245b732d9fffb1397