linux-ramdump-parser-v2: Parse socinfo only if needed
If user specified --force-hardware, then skip the socinfo parsing to
determine the hardware/board information. The user is warned about
using the force-hardware option any way.

Checking for valid swapper_pg_dir in the vmlinux also ensures that
the parsing doesn't trip over when --force-hardware is used along with
a bad or stripped vmlinux.

Change-Id: Icce1556b39a7533418668a7f38af30f4463c32e5