linux-ramdump-parser-v2: Optimize taskdump panic stack dumping
When attempting to locate panicking tasks via the
"--check-for-panic" option, the taskdump parser currently attempts
to exhaustively check every task in kernel task list to see if it
panicked.

To improve performance, task checking can be limited to tasks that
were listed as currently running at the time the RAM dump was
collected. Other tasks do not need to be checked because a
panicking process is expected to disable IRQs and preemption after
entering the panic handler, and therefore will not yield to any
other tasks executing on the same CPU prior to resetting.

Change-Id: I51d7b979f918e21bfdd2adf2a3429883f11bab6a