ipc_logging: Update "test" command to accept version, add --64-bit option
When processing a log page, the page header size depends on the size of
internal structures and pointers that can vary with machine size
(32-bit or 64-bit). This information is stored in the logging context
during automated extraction (i.e. when the "parse" command is used), but in
the "test" command only the log pages are provided, so the necessary
information is not present.

The correct page header size can be determined if the version of IPC
Logging and the architecture size (32-bit or 64-bit) are known. Store
the possible header sizes as module variables, modify the test command
to accept the IPC Logging version, and modify the script to take an
argument denoting if the logs are interpreted as 32-bit or 64-bit.
In the "test" command, select the correct header size using the IPC
Logging version and architecture size.

Change-Id: I53ccdc7982b5cfd79dfb6c36d982037b3b2a3955