ipc_logging: Update logging, output directory configurations
It is necessary to allow the script to get included as a module by other
scripts. Currently, the logging and output directory configuration are
performed in the script's main function, and the options from parsing
the command line are not passed from function to function.

Create two new functions for the logging and output directory
configuraiton, and pass the command-line options object to each
function that uses it. This allows the options to be passed in from
other scripts.

Change-Id: Ibf34d72cd70620bcbcc00e48141a87a721b1f987