linux-ramdump-parser-v2: Pass options object to RamDump constructor
The constructor currently used in ramparse.py to initialize the RamDump
object has a very long argument list. This list can be shortened by
passing the options object to the constructor, rather than enumerating
every member of the options object as a separate argument.

Pass the options object to the RamDump constructor instead of passing
each member as a separate argument.

Change-Id: I7cbf173232838dd8f9fdbdad0e2b7336ad8fc07b