ipc_logging: Handle corrupted log page headers
When parsing IPC logs, sometimes a bogus log page header is present which
overwrites the bit used to discern whether or not the log page is version
0. This causes the script to read the version as being 1 or greater, and
the script then asserts after failing to find a log context for the page.
In general, header corruption like this may cause other issues that could
lead to the context not being found, which currently aborts the script.

Modify the script to continue parsing the logs as logs of an unknown
version in this case. If the header is valid, the proper version will be
parsed from it, allowing parsing to continue normally.

Change-Id: I02ae617980f817705224b7b02926a9f576dd88e6