lrdpv2: Fix an infinite loop while iterating RT data structure
Ram dump parser compares RT data structure list head location
with first element in the list to determine whether list is
empty. However it never checks for cycle in the list and runs
forever in cases where cycle exists in the list. Fix this by
tracking the elements seen and break the loop when cycle is
found.

Change-Id: I971c59b81bf340eddff6c8225aebb163af20aba8