ipc_logging: Add support for parsing QTimer timestamp
IPC Logging currently displays the kernel log timestamp, which is based
on a timer that does not increase when the Apps processor is in suspend.
This means that the kernel log timestamp "loses time" and cannot be
correlated with events on other processors. The source of the QTimer
timestamp is always running and has the same source other logs in the
system, which makes it convenient to correlate logs across the system.

Update the IPC Logging extraction script to handle version 2 of IPC
Logging, which includes the QTimer timestamp.

Change-Id: Ia4534a53012e29efe5ac647b6212530f84c07f46