ipc_logging: Fix QTimer packing functions
Along with the unpack() functions which are present for parsing, the
classes which parse individual log messages also have pack() functions
which are used for recreating the log messages in their binary form.
These functions are used for unit testing.

In the TSVQTimer pack() function, incorrect math is performed on the
result before it is returned. In the LogMessageString pack() function, a
QTimer object is manipulated before it is created.

Fix these errors.

Change-Id: I5c288ddd53e468368c3663bdc9c1134a5f694865